Add feedback "tones"
Feedback items can now be created with a "tone" – that is, whether the
feedback is positive, negative, or passive. The parameter is optional
since not all feedback will have a tone associated with it.

Feedback with positive, passive, or negative tones will post to
Hipchat with the colors, green, yellow, and red, respectively. Neutral
feedback items will still be gray.